refact: components organized in class folder
