+++ components/cart/Cart.tsx
@@ -1,4 +1,4 @@
-import { CartLogo } from '../ui/icons';
+import { CartLogo } from '../../ui/icons';
 import { Button } from 'ui/button/styled';
 import { useCart, useCleanCart, useMe, useTotalCart } from 'hooks';
 import { useEffect, useState } from 'react';
