+++ components/cart/ItemCounter.tsx
