+++ components/cart/Orders.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { ButtonDanger, ButtonPrimary } from 'ui/button/styled';
 import { LabelPrimary } from 'ui/label/styled';
-import { Link } from '../ui/Link';
-import { ContainerOrder, ItemLinkOrder, ItemOrder } from '../ui/orders/styled';
+import { Link } from '../../ui/Link';
+import { ContainerOrder, ItemLinkOrder, ItemOrder } from '../../ui/orders/styled';
 
 export const Order = ({
   index,
