+++ components/index.ts
@@ -1,10 +1,10 @@
-export * from "./Header";
-export * from "./LoginCode";
-export * from "./LoginEmail";
-export * from "./Product";
-export * from "./ProductFeatured";
-export * from "./Pagination";
-export * from "./Searcher";
-export * from "./OrdersMobile";
-export * from "./Orders";
-export * from "./MyTextarea";
+export * from './ui/Header';
+export * from './auth/LoginCode';
+export * from './auth/LoginEmail';
+export * from './product/Product';
+export * from './product/ProductFeatured';
+export * from './ui/Pagination';
+export * from './ui/Searcher';
+export * from './cart/OrdersMobile';
+export * from './cart/Orders';
+export * from './ui/MyTextarea';
