+++ package.json
@@ -26,7 +26,7 @@
     "autoprefixer": "^10.4.13",
     "axios": "^1.1.3",
     "cypress": "^12.0.1",
-    "daisyui": "^2.40.1",
+    "daisyui": "^2.51.5",
     "dotenv": "^16.0.3",
     "eslint": "8.27.0",
     "eslint-config-next": "13.0.3",
