+++ pages/_document.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import Document, { Html, Head, Main, NextScript } from 'next/document';
 import { CssBaseline } from '@nextui-org/react';
-import Script from 'next/script';
 
 class MyDocument extends Document {
   static async getInitialProps(ctx: any) {
