+++ ui/index.ts
@@ -1,7 +1,7 @@
 export * from "./banner/Banner";
 export * from "./Payment";
 export * from "./wrappers/User";
-export * from "../components/Cart";
+export * from "../components/cart/Cart";
 export * from "./wrappers/Featured";
 export * from "./wrappers/Categories";
 export * from "./button/styled";
@@ -11,7 +11,7 @@ export * from "./field/Field";
 export * from "./footer/Footer";
 export * from "./Link";
 export * from "./loaders/Loader";
-export * from "../components/Orders";
+export * from "../components/cart/Orders";
 export * from "./Toast";
 
 export * from "./wrappers/Basic";
